Refactor - remove duplicate AdditionalAttackPhaseEffect classes
